My brain is fried, it took me a long time to figure out that value = {value} meant that the {value} was the various States.  HINT use the Spread Operator on it!!!!!
Then using the SPREAD OPERATOR... we could extract the individual STATES we wanted to use!

Then refactoring the code, we needed to remove the existing variable that was being passed in the function.

Once the variable was removed we needed to introduce the variable using the CONTEXT method of Passing Data Deeply.